requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
deprecation==2.1.0
fastapi==0.115.13
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
gotrue==2.12.3
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
Levenshtein==0.27.1
lxml==5.4.0
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
numpy==2.3.1
packaging==25.0
pillow==11.2.1
playwright==1.53.0
postgrest==1.0.2
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.13.0
realtime==2.4.3
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.3
supafunc==0.9.4
sympy==1.14.0
thefuzz==0.22.1
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
websockets==14.2
yarl==1.20.1
